Create Ezekiel Mohale
Mr Mohale has a burning desire and need for wanting to partner with government to work for the betterment of society. However, he points out and would really love to drive the point home that this non - profit Youth Organisation is not involved in party politics, even though there is a great need to work with the government towards creating more economic and upliftment opportunities for our people. 

Even where there are visible failures on the part of government entities to uplift our communities out of their underdevelopment, the idea for Mr Mohale and the team is not to judge the government harshly, but to work with them to create opportunities for development and self-determination.

The BRT team is passionate to deal with challenges such as are faced by our unemployed black graduates who are struggling to find employment in the state and private sectors. This organisation wants to help find solutions that will help our people improve their lives and change their circumstances for the better. And further admit that the organization does have those solutions in black and white to provide to the nation. We can assist our government with ideas that can help our country move forward.